Change project layout
Since I discoverred that windows does not have 'exec' in the same way as
unix does, the only way to shim is to keep the shim running as long as
the child command runs. This makes the idea of a shim calling 'asdfw
exec ...' invalid as we'll have double wrappers. all the logic of
getting the version to run should exist in the shim itself. Because of
that there's no point in two separates projects.